Combine HiveCatalog and GenericHiveMetastoreCatalog
