+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/catalog/GenericCatalogTable.java
@@ -26,73 +26,43 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static org.apache.flink.util.Preconditions.checkNotNull;
-
 /**
  * A generic catalog table implementation.
  */
-public class GenericCatalogTable implements CatalogTable {
-	// Schema of the table (column names and types)
-	private final TableSchema tableSchema;
-	// Partition keys if this is a partitioned table. It's an empty set if the table is not partitioned
-	private final List<String> partitionKeys;
-	// Properties of the table
-	private final Map<String, String> properties;
-	// Comment of the table
-	private final String comment;
+public class GenericCatalogTable extends AbstractCatalogTable {
+
+	public GenericCatalogTable(
+		TableSchema tableSchema,
+		Map<String, String> properties,
+		String comment) {
+		this(tableSchema, new ArrayList<>(), properties, comment);
+	}
 
 	public GenericCatalogTable(
 			TableSchema tableSchema,
 			List<String> partitionKeys,
 			Map<String, String> properties,
 			String comment) {
-		this.tableSchema = checkNotNull(tableSchema, "tableSchema cannot be null");
-		this.partitionKeys = checkNotNull(partitionKeys, "partitionKeys cannot be null");
-		this.properties = checkNotNull(properties, "properties cannot be null");
-		this.comment = comment;
-	}
-
-	public GenericCatalogTable(
-			TableSchema tableSchema,
-			Map<String, String> properties,
-			String description) {
-		this(tableSchema, new ArrayList<>(), properties, description);
-	}
-
-	@Override
-	public boolean isPartitioned() {
-		return !partitionKeys.isEmpty();
-	}
-
-	@Override
-	public List<String> getPartitionKeys() {
-		return partitionKeys;
-	}
-
-	@Override
-	public Map<String, String> getProperties() {
-		return properties;
-	}
-
-	@Override
-	public TableSchema getSchema() {
-		return this.tableSchema;
-	}
-
-	@Override
-	public String getComment() {
-		return comment;
+		super(tableSchema, partitionKeys, properties, comment);
+		properties.put(GenericInMemoryCatalog.FLINK_META_PROPERTY_KEY, GenericInMemoryCatalog.FLINK_META_PROPERTY_VALUE);
 	}
 
 	@Override
 	public GenericCatalogTable copy() {
 		return new GenericCatalogTable(
-			this.tableSchema.copy(), new ArrayList<>(partitionKeys), new HashMap<>(this.properties), comment);
+			getSchema().copy(), new ArrayList<>(getPartitionKeys()), new HashMap<>(getProperties()), getComment());
+	}
+
+	@Override
+	public Map<String, String> toProperties() {
+		// TODO: Filter out ANY properties that are not needed for table discovery.
+		Map<String, String> properties = new HashMap<>();
+		return properties;
 	}
 
 	@Override
 	public Optional<String> getDescription() {
-		return Optional.of(comment);
+		return Optional.of(getComment());
 	}
 
 	@Override
